Adding internal errors helper file
